T7 Move onResize listener to a custom hook
